Estrutura geral da aplicação e implementação inicial da interface gráfica e das operações.

+++ src/main/java/br/com/paygo/enums/PWOper.java
@@ -4,17 +4,17 @@ package br.com.paygo.enums;
  * Tipos de operações disponibilizadas
  */
 public enum PWOper {
-    ADMIN(32),      // Qualquer transação que não seja um pagamento(estorno, pré-autorização, consulta, relatório, reimpressão de recibo, etc.).
-    SALE(33),       // Pagamento de mercadorias ou serviços.
-    SALEVOID(34);   // Estorna uma transação de venda que foi previamente realizada e confirmada
+    ADMIN((short)32),      // Qualquer transação que não seja um pagamento(estorno, pré-autorização, consulta, relatório, reimpressão de recibo, etc.).
+    SALE((short)33),       // Pagamento de mercadorias ou serviços.
+    SALEVOID((short)34);   // Estorna uma transação de venda que foi previamente realizada e confirmada
 
-    private final int value;
+    private final short value;
 
-    PWOper(int value) {
+    PWOper(short value) {
         this.value = value;
     }
 
-    public int getValue() {
+    public short getValue() {
         return value;
     }
 }
